@@ -1,31 +1,28 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 29
-    buildToolsVersion "29.0.3"
+    compileSdkVersion 32
 
     defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 29
-        versionCode 5
-        versionName "1.3.1-SNAPSHOT"
+        minSdkVersion 16
+        targetSdkVersion 32
     }
     buildTypes {
         release {
             minifyEnabled false
         }
     }
+    namespace 'cat.ereza.logcatreporter'
 }
 
 ext {
     PUBLISH_GROUP_ID = 'cat.ereza'
     PUBLISH_ARTIFACT_ID = 'logcatreporter'
-    PUBLISH_VERSION = android.defaultConfig.versionName
+    PUBLISH_VERSION = '1.4.0'
 }
 
 apply from: "publish-mavencentral.gradle"
 
 dependencies {
-    api 'com.crashlytics.sdk.android:crashlytics:2.10.1'
-    implementation 'androidx.annotation:annotation:1.1.0'
+    api 'com.google.firebase:firebase-crashlytics:18.2.11'
 }
