@@ -1,20 +1,15 @@
 apply plugin: 'com.android.application'
-apply plugin: 'io.fabric'
-
-repositories {
-    maven { url 'https://maven.fabric.io/public' }
-}
+apply plugin: 'com.google.gms.google-services'
+apply plugin: 'com.google.firebase.crashlytics'
 
 android {
-    compileSdkVersion 29
-    buildToolsVersion "29.0.3"
+    compileSdkVersion 32
 
     defaultConfig {
-        applicationId "cat.ereza.logcatreporter.sample"
-        minSdkVersion 14
-        targetSdkVersion 29
+        minSdkVersion 16
+        targetSdkVersion 32
         versionCode 5
-        versionName "1.3.1-SNAPSHOT"
+        versionName "1.4.0"
     }
     buildTypes {
         debug {
@@ -26,6 +21,7 @@ android {
             proguardFiles 'proguard-rules.pro'
         }
     }
+    namespace "cat.ereza.logcatreporter.sample"
 }
 
 dependencies {
